moved common classes to __init__.py
